rename getdate file to helperFunc and update the imports
